backends/nemotron-parse/pyproject.toml
[project]
name = "nemotron-parse-service"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "pillow>=10.4.0",
    "uvicorn[standard]>=0.32.0",
    "transformers==4.51.3",
    "accelerate==1.12.0",
    "albumentations==2.0.8",
    "timm==1.0.22",
    "open-clip-torch>=2.29.0",
    "einops>=0.8.0",
    "torch>=2.0.0",
    "beautifulsoup4>=4.12.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
